backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
web3
qiskit
pennylane
pillow
numpy
scikit-learn
transformers
torch
opencv-python
scipy
opencv-contrib-python
tensorflow
librosa
soundfile
pennylane